Verify simulation amount
- Extended macro to check if asset is gas_asset
- Added test to verify simulation amounts for none gas assets
- Moved get_simulation_amount to ingress-egress pallet
